chatbot/requirements.txt
flask
transformers
wikipedia-api
